fix: Use Web Page Block for Navbar and Footer
- Add more Web Templates
