+++ frappe/templates/components/web_blocks.html
@@ -1,0 +1,3 @@
+{%- for web_block in web_blocks -%}
+{{ c('web_block', web_block=web_block, htmltag=htmltag) }}
+{%- endfor -%}
