--- frappe/templates/components/web_page_section.html
@@ -1,5 +1,0 @@
-<section class="py-32 {{ section.css_class or '' }}" data-section-title="{{ section.title | e }}">
-    <div class="container">
-        {{ section.rendered_html }}
-    </div>
-</section>
