+++ frappe/website/doctype/web_page/web_page.js
@@ -2,6 +2,9 @@
 // MIT License. See license.txt
 
 frappe.ui.form.on('Web Page', {
+	onload: function() {
+		frappe.require('/assets/frappe/js/frappe/utils/web_page_block.js');
+	},
 	title: function(frm) {
 		if (frm.doc.title && !frm.doc.route) {
 			frm.set_value('route', frappe.scrub(frm.doc.title, '-'));
@@ -43,29 +46,5 @@ frappe.ui.form.on('Web Page', {
 
 	set_meta_tags(frm) {
 		frappe.utils.set_meta_tag(frm.doc.route);
-	}
-});
-
-frappe.ui.form.on('Web Page Block', {
-	edit_values(frm, cdt, cdn) {
-		let row = frm.selected_doc;
-		frappe.model.with_doc('Web Template', row.web_template).then(doc => {
-			let d = new frappe.ui.Dialog({
-				title: __('Edit Values'),
-				fields: doc.fields,
-				primary_action(values) {
-					frappe.model.set_value(
-						cdt,
-						cdn,
-						'web_template_values',
-						JSON.stringify(values)
-					);
-					d.hide();
-				}
-			});
-			let values = JSON.parse(row.web_template_values || '{}');
-			d.set_values(values);
-			d.show();
-		});
 	}
 });
