+++ frappe/website/doctype/web_page_block/web_page_block.json
@@ -8,7 +8,11 @@
   "web_template",
   "edit_values",
   "web_template_values",
-  "css_class"
+  "css_class",
+  "column_break_5",
+  "add_container",
+  "add_padding",
+  "add_shade"
  ],
  "fields": [
   {
@@ -27,6 +31,7 @@
   {
    "fieldname": "web_template_values",
    "fieldtype": "Code",
+   "hidden": 1,
    "label": "Web Template Values",
    "options": "JSON"
   },
@@ -34,11 +39,33 @@
    "fieldname": "css_class",
    "fieldtype": "Small Text",
    "label": "CSS Class"
+  },
+  {
+   "fieldname": "column_break_5",
+   "fieldtype": "Column Break"
+  },
+  {
+   "default": "1",
+   "fieldname": "add_padding",
+   "fieldtype": "Check",
+   "label": "Add Padding"
+  },
+  {
+   "default": "0",
+   "fieldname": "add_shade",
+   "fieldtype": "Check",
+   "label": "Shaded Section"
+  },
+  {
+   "default": "1",
+   "fieldname": "add_container",
+   "fieldtype": "Check",
+   "label": "Add Container"
   }
  ],
  "istable": 1,
  "links": [],
- "modified": "2020-04-18 12:50:00.070761",
+ "modified": "2020-04-19 16:16:44.524042",
  "modified_by": "Administrator",
  "module": "Website",
  "name": "Web Page Block",
