+++ frappe/website/doctype/web_template_field/web_template_field.json
@@ -7,7 +7,8 @@
  "field_order": [
   "label",
   "fieldname",
-  "fieldtype"
+  "fieldtype",
+  "reqd"
  ],
  "fields": [
   {
@@ -28,13 +29,19 @@
    "fieldtype": "Select",
    "in_list_view": 1,
    "label": "Fieldtype",
-   "options": "Data\nSmall Text\nText\nAttach Image",
+   "options": "Data\nSmall Text\nText\nAttach Image\nMarkdown Editor",
    "reqd": 1
+  },
+  {
+   "default": "0",
+   "fieldname": "reqd",
+   "fieldtype": "Check",
+   "label": "Mandatory"
   }
  ],
  "istable": 1,
  "links": [],
- "modified": "2020-04-17 19:30:33.829203",
+ "modified": "2020-04-19 15:56:01.606898",
  "modified_by": "Administrator",
  "module": "Website",
  "name": "Web Template Field",
