+++ frappe/website/web_template/footer_horizontal/footer_horizontal.json
@@ -1,0 +1,97 @@
+{
+ "creation": "2020-04-19 15:01:23.344637",
+ "docstatus": 0,
+ "doctype": "Web Template",
+ "fields": [
+  {
+   "fieldname": "brand_image",
+   "fieldtype": "Attach Image",
+   "label": "Brand Image",
+   "reqd": 0
+  },
+  {
+   "fieldname": "brand_label",
+   "fieldtype": "Data",
+   "label": "Brand Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "address",
+   "fieldtype": "Small Text",
+   "label": "Address",
+   "reqd": 0
+  },
+  {
+   "fieldname": "copyright",
+   "fieldtype": "Small Text",
+   "label": "Copyright",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_1_label",
+   "fieldtype": "Data",
+   "label": "Item 1 Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_1_url",
+   "fieldtype": "Data",
+   "label": "Item 1 URL",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_2_label",
+   "fieldtype": "Data",
+   "label": "Item 2 Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_2_url",
+   "fieldtype": "Data",
+   "label": "Item 2 URL",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_3_label",
+   "fieldtype": "Data",
+   "label": "Item 3 Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_3_url",
+   "fieldtype": "Data",
+   "label": "Item 3 URL",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_4_label",
+   "fieldtype": "Data",
+   "label": "Item 4 Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_4_url",
+   "fieldtype": "Data",
+   "label": "Item 4 URL",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_5_label",
+   "fieldtype": "Data",
+   "label": "Item 5 Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_5_url",
+   "fieldtype": "Data",
+   "label": "Item 5 URL",
+   "reqd": 0
+  }
+ ],
+ "idx": 0,
+ "modified": "2020-04-19 15:11:42.097670",
+ "modified_by": "Administrator",
+ "name": "Footer Horizontal",
+ "owner": "Administrator",
+ "standard": 1
+}
