+++ frappe/website/web_template/hero/hero.html
@@ -1,0 +1,31 @@
+<div>
+	<h2 class="text-4xl font-extrabold leading-10 tracking-tight text-black sm:text-5xl sm:leading-none md:text-5xl">
+		{{ title }}
+	</h2>
+	{%- if subtitle -%}
+	<p
+		class="max-w-2xl mt-3 text-base text-gray-800 sm:mt-5 sm:text-lg md:mt-5 md:text-xl lg:mx-0">
+		{{ subtitle }}
+	</p>
+	{%- endif -%}
+	{%- if primary_action or secondary_action -%}
+	<div class="mt-5 sm:mt-8 sm:flex sm:justify-center lg:justify-start">
+		{%- if primary_action -%}
+		<div class="rounded-md shadow">
+			<a href="{{ primary_action }}"
+				class="flex items-center justify-center w-full px-8 py-3 text-base font-medium leading-6 text-white transition duration-150 ease-in-out bg-blue-500 border border-transparent rounded-md hover:bg-blue-400 focus:outline-none focus:shadow-outline md:py-4 md:text-lg md:px-10">
+				{{ primary_action_label }}
+			</a>
+		</div>
+		{%- endif -%}
+		{%- if secondary_action -%}
+		<div class="mt-3 sm:mt-0 sm:ml-3">
+			<a href="{{ secondary_action }}"
+				class="flex items-center justify-center w-full px-8 py-3 text-base font-medium leading-6 text-blue-700 transition duration-150 ease-in-out bg-blue-100 border border-transparent rounded-md hover:text-blue-600 hover:bg-blue-50 focus:outline-none focus:shadow-outline focus:border-blue-300 md:py-4 md:text-lg md:px-10">
+				{{ secondary_action_label }}
+			</a>
+		</div>
+		{%- endif -%}
+	</div>
+	{%- endif -%}
+</div>
