+++ frappe/website/web_template/markdown/markdown.html
@@ -1,0 +1,3 @@
+<div class="max-w-2xl from-markdown">
+	{{ frappe.utils.md_to_html(content) }}
+</div>
