+++ frappe/website/web_template/markdown/markdown.json
@@ -1,0 +1,19 @@
+{
+ "creation": "2020-04-19 15:56:20.833205",
+ "docstatus": 0,
+ "doctype": "Web Template",
+ "fields": [
+  {
+   "fieldname": "content",
+   "fieldtype": "Markdown Editor",
+   "label": "Content",
+   "reqd": 0
+  }
+ ],
+ "idx": 0,
+ "modified": "2020-04-19 15:56:20.833205",
+ "modified_by": "Administrator",
+ "name": "Markdown",
+ "owner": "Administrator",
+ "standard": 1
+}
