+++ frappe/website/web_template/navbar_with_links_on_right/navbar_with_links_on_right.html
@@ -1,0 +1,26 @@
+<header class="px-4 py-4 bg-white border-b border-gray-200 sm:px-6 lg:px-8">
+	<div
+		class="container relative z-10 flex items-center justify-between">
+		<div class="flex items-end w-24">
+			<a href="/" class="block">
+				<img class="w-auto h-8 sm:h-10" src="{{ brand_image }}" alt="Brand Logo">
+			</a>
+		</div>
+		<div class="flex items-center text-sm leading-5 space-x-8">
+			{%- for index in ['1', '2', '3', '4', '5'] -%}
+
+			{%- set label = values['item_' + index + '_label'] -%}
+			{%- set url = values['item_' + index + '_url'] -%}
+
+			{%- if label and url -%}
+			{{ c('navbar_link', label=label, url=url) }}
+			{%- endif -%}
+
+			{%- endfor -%}
+
+			{%- if cta_label and cta_url -%}
+			{{ c('button', label=cta_label, url=cta_url, variant="primary") }}
+			{%- endif -%}
+		</div>
+	</div>
+</header>
