+++ frappe/website/web_template/navbar_with_links_on_right/navbar_with_links_on_right.json
@@ -1,0 +1,97 @@
+{
+ "creation": "2020-04-19 13:49:33.971237",
+ "docstatus": 0,
+ "doctype": "Web Template",
+ "fields": [
+  {
+   "fieldname": "brand_image",
+   "fieldtype": "Attach Image",
+   "label": "Brand Image",
+   "reqd": 0
+  },
+  {
+   "fieldname": "brand_label",
+   "fieldtype": "Data",
+   "label": "Brand Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "cta_label",
+   "fieldtype": "Data",
+   "label": "CTA Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "cta_url",
+   "fieldtype": "Data",
+   "label": "CTA URL",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_1_label",
+   "fieldtype": "Data",
+   "label": "Item 1 Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_1_url",
+   "fieldtype": "Data",
+   "label": "Item 1 URL",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_2_label",
+   "fieldtype": "Data",
+   "label": "Item 2 Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_2_url",
+   "fieldtype": "Data",
+   "label": "Item 2 URL",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_3_label",
+   "fieldtype": "Data",
+   "label": "Item 3 Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_3_url",
+   "fieldtype": "Data",
+   "label": "Item 3 URL",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_4_label",
+   "fieldtype": "Data",
+   "label": "Item 4 Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_4_url",
+   "fieldtype": "Data",
+   "label": "Item 4 URL",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_5_label",
+   "fieldtype": "Data",
+   "label": "Item 5 Label",
+   "reqd": 0
+  },
+  {
+   "fieldname": "item_5_url",
+   "fieldtype": "Data",
+   "label": "Item 5 URL",
+   "reqd": 0
+  }
+ ],
+ "idx": 0,
+ "modified": "2020-04-19 13:51:26.391098",
+ "modified_by": "Administrator",
+ "name": "Navbar with Links on Right",
+ "owner": "Administrator",
+ "standard": 1
+}
