+++ frappe/website/web_template/section_with_big_cards/section_with_big_cards.json
@@ -6,51 +6,84 @@
   {
    "fieldname": "title",
    "fieldtype": "Data",
-   "label": "Title"
+   "label": "Title",
+   "reqd": 0
   },
   {
    "fieldname": "card_1_title",
    "fieldtype": "Data",
-   "label": "Card 1 Title"
+   "label": "Card 1 Title",
+   "reqd": 0
   },
   {
    "fieldname": "card_1_content",
    "fieldtype": "Text",
-   "label": "Card 1 Content"
+   "label": "Card 1 Content",
+   "reqd": 0
+  },
+  {
+   "fieldname": "card_1_url",
+   "fieldtype": "Data",
+   "label": "Card 1 URL",
+   "reqd": 0
   },
   {
    "fieldname": "card_2_title",
    "fieldtype": "Data",
-   "label": "Card 2 Title"
+   "label": "Card 2 Title",
+   "reqd": 0
   },
   {
    "fieldname": "card_2_content",
    "fieldtype": "Text",
-   "label": "Card 2 Content"
+   "label": "Card 2 Content",
+   "reqd": 0
+  },
+  {
+   "fieldname": "card_2_url",
+   "fieldtype": "Data",
+   "label": "Card 2 URL",
+   "reqd": 0
   },
   {
    "fieldname": "card_3_title",
    "fieldtype": "Data",
-   "label": "Card 3 Title"
+   "label": "Card 3 Title",
+   "reqd": 0
   },
   {
    "fieldname": "card_3_content",
    "fieldtype": "Text",
-   "label": "Card 3 Content"
+   "label": "Card 3 Content",
+   "reqd": 0
+  },
+  {
+   "fieldname": "card_3_url",
+   "fieldtype": "Data",
+   "label": "Card 3 URL",
+   "reqd": 0
   },
   {
    "fieldname": "card_4_title",
    "fieldtype": "Data",
-   "label": "Card 4 Title"
+   "label": "Card 4 Title",
+   "reqd": 0
   },
   {
    "fieldname": "card_4_content",
    "fieldtype": "Text",
-   "label": "Card 4 Content"
+   "label": "Card 4 Content",
+   "reqd": 0
+  },
+  {
+   "fieldname": "card_4_url",
+   "fieldtype": "Data",
+   "label": "Card 4 URL",
+   "reqd": 0
   }
  ],
  "idx": 0,
- "modified": "2020-04-17 19:42:19.712813",
+ "modified": "2020-04-19 15:57:56.743514",
  "modified_by": "Administrator",
  "name": "Section with Big Cards",
  "owner": "Administrator",
